Get the Node PubKey from a BindMessage dogenet sends us
And profile fields validation in web API

Re-using the stored announcement is broken, because it typically
includes the Node PubKey we had last time we were running.
But now we've just started up and don't have one.

Loading the announce needs to be deferred until we have the Node PubKey
or we need to store and expire node pubkeys in our database.