Store the list of associated nodes for the identity profile
When we connect to DogeNet as a protocol handler, we receive the
public key of the node. This pubkey needs to be included in our
signed identity message to prove that the identity claims the node
and the node claims the identity.

We store the node pubkey(s) in the database so, on startup, we have
them on-hand when we make the identity message. This allows us to
compare and re-use the previously signed message without generating
a new signature, which would cause spurious gossip churn.